расстановка индексов, раздел оптимизации бд в админке,

+++ protected/modules/admin/controllers/DbController.php
@@ -10,20 +10,22 @@ class DbController extends Controller
 {
     public function actionBackup()
     {
-        $now = date("_Y-n-d__H-i-s");
-        $filename = "BACKUP_ORTHO_DB_" . $now . ".sql";
-        $command = "mysqldump --flush-logs --lock-tables --databases -u" . Yii::app()->db->username . " -p" . Yii::app()->db->password . " -hlocalhost ortho_db > $filename";
-        $result = system($command);
-        if (!$result) {
-            header("Content-Type: text/plain;");
-            header('Content-Description: File Transfer');
-            header('Content-Type: application/octet-stream');
-            header('Content-Transfer-Encoding: binary');
-            header("Content-Disposition: attachment; filename='$filename'");
-            echo file_get_contents($filename);
-            unlink($filename);
-            Yii::app()->end();
-        }
+        if (Yii::app()->request->isPostRequest && isset($_POST['backupDbBtn'])) {
+            $now = date("_Y-n-d__H-i-s");
+            $filename = "BACKUP_ORTHO_DB_" . $now . ".sql";
+            $command = "mysqldump --flush-logs --lock-tables --databases -u" . Yii::app()->db->username . " -p" . Yii::app()->db->password . " -hlocalhost ortho_db > $filename";
+            $result = system($command);
+            if (!$result) {
+                header("Content-Type: text/plain;");
+                header('Content-Description: File Transfer');
+                header('Content-Type: application/octet-stream');
+                header('Content-Transfer-Encoding: binary');
+                header("Content-Disposition: attachment; filename='$filename'");
+                echo file_get_contents($filename);
+                unlink($filename);
+            }
+        } else
+            $this->render('backup');
     }
 
     public function actionRestore()
@@ -48,6 +50,23 @@ class DbController extends Controller
         $this->render('restore', [
             'model' => $model,
         ]);
+    }
+
+    public function actionOptimize()
+    {
+        if (Yii::app()->request->isAjaxRequest) {
+            $connection = Yii::app()->db;
+            $tables = $connection->createCommand("SELECT table_name
+                FROM information_schema.tables
+                WHERE table_schema='ortho_db' AND table_name <> 'tbl_migration'")
+                ->queryAll();
+
+            foreach ($tables as $table) {
+                $connection->createCommand('OPTIMIZE TABLE ' . $table['table_name'])->execute();
+            }
+            Yii::app()->end();
+        } else
+            $this->render('optimize');
     }
 
 }
